Fix code review issue
According to the project requirements, each of the individual classes should be in a different file not in the same file as the Nameable class.
