+++ persons.rb
@@ -1,5 +1,7 @@
 require 'securerandom'
 require './nameable'
+require './capitalize_decorator'
+require './trimmer_decorator'
 class Person < Nameable
   # creating accessors
   attr_accessor :name, :age
@@ -31,3 +33,10 @@ class Person < Nameable
     false
   end
 end
+
+person = Person.new('maximilianus', 22)
+pp person.correct_name
+capitalizedPerson = CapitalizeDecorator.new(person)
+pp capitalizedPerson.correct_name
+capitalizedTrimmedPerson = TrimmerDecorator.new(capitalizedPerson)
+pp capitalizedTrimmedPerson.correct_name
