+++ trimmer_decorator.rb
@@ -1,0 +1,9 @@
+require './decorator'
+
+class TrimmerDecorator < Decorator
+  def correct_name
+    return @nameable.correct_name[0..9] if @nameable.correct_name.length > 10
+
+    @nameable.correct_name
+  end
+end
